Update to new arrow avro API
Also reuse buffers in writing avro.
This will improve performance when writing multiple
chunks
