+++ polars/polars-core/Cargo.toml
@@ -173,10 +173,11 @@ thiserror = "^1.0"
 
 [dependencies.arrow]
 package = "arrow2"
-git = "https://github.com/ritchie46/arrow2"
-# rev = "5711cfcbb785450697eb38e221b5dc1825550ee6"
+git = "https://github.com/jorgecarleitao/arrow2"
+# git = "https://github.com/ritchie46/arrow2"
+rev = "3b9d86bfb65bc786bf9315a41737a77895716443"
 # path = "../../../arrow2"
-branch = "ipc_meta"
+# branch = "ipc_meta"
 # version = "0.12"
 default-features = false
 features = [
